requirements.txt
matplotlib==3.1.0
dask==2.6.0
imutils==0.5.3
numpy==1.16.4
tensorflow==1.14.0
Keras==2.2.4
tqdm==4.32.2
Pillow==6.2.1
scikit_learn==0.22
